Worked on reducing fluff in Spiking Network main.
There was a lot (and still is) of cross over between the multichannel and single channel model. I am working on removing all the multichannel stuff for the single channel specific things. Eventually we will have clean commented organized scripts and we can put a bow on the single channel effort.