Updating Gradients to be wrt parameters
In the last week or so I have managed to put the gradient updates in the forwards loop. Doing so I was able to make it so that the PSC updates can be involved. Just wanna save a version here where the old grads and the new grads are still avaiable.